% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctwas_weights_poisson.R
\name{compute_weights_poisson}
\alias{compute_weights_poisson}
\title{Train model to predict genetically regulated gene expression using single-cell eQTL data}
\usage{
compute_weights_poisson(
  y,
  geno_cell,
  pt,
  knots = c(0.25, 0.5, 0.75),
  degree = 3,
  lambda = NULL,
  nlambda = 50,
  libsize = NULL,
  covar = NULL
)
}
\arguments{
\item{y}{Expression count data for one gene. Integer vector of length ncells (number of cells).}

\item{geno_cell}{Genotype of cis-SNPs of the gene. Matrix of dimensions ncells x nsnps (number of SNPs). Each row is a cell and each column is a SNP. Although SNP values are the same across cells within the same individual, they are repeated for each cell in \code{geno_cell}.}

\item{pt}{Pseudotime of the cells. Pseudotime values learned from packages such as Slingshot, TSCAN should be scaled to rank/(number of cells) such that values are approximately uniformly distributed between 0 and 1.}

\item{knots}{Internal knots for B-spline basis functions that model SNP effects on gene expression. Not including 0 and 1. Default to be \code{c(0.25,0.5,0.75)}.}

\item{degree}{Degree of B-spline basis functions, default to 3 (cubic B-spline).}

\item{lambda}{Tuning parameter for group lasso penalty. If null, `lambda` is selected by cross-validation.}

\item{nlambda}{Number of `lambda` values in cross-validation - default is 50.}

\item{libsize}{Library size of the cells. Vector of length ncells.}

\item{covar}{Covariate matrix to be adjusted. Not penalized.}
}
\value{
An object of class \code{grpreg} from the \code{grpreg} package, with the following additional entries:
\item{Wmat}{Coefficients representing effects of SNPs on gene expression. Matrix of (number of SNPs) x (number of B-spline basis functions) generated by reshaping \code{beta}. The effect of SNP k on gene expression over pseudotime t is \eqn{w_k(t)=\sum_m w_{km}\phi_m(t)}, where \eqn{w_{km}} is the \code{(k,m)} entry of \code{Wmat}.}
\item{knots}{Internal knots for B-spline basis functions}
\item{degree}{Degree of B-spline basis functions.}
\item{lambda.seq}{Number of \code{lambda}'s to test in cross-validation.}
Other entries are \code{beta}, \code{family}, \code{group}, \code{lambda}, \code{alpha}, \code{deviance}, \code{linear.predictors}, \code{n}, \code{penalty}, \code{df}, \code{iter}, \code{group.multiplier}, \code{y}. See \code{grpreg} package (https://cran.r-project.org/web/packages/grpreg/index.html) for details.
}
\description{
In TWiST, prediction models are trained using scalar-on-function regression with cis-SNPs as predictors and gene expression counts from single-cell RNA-seq as outcomes.
}
\details{
Additional details...
}
\examples{
library(DAESC)
data("example", package="DAESC")
res.bb <- daesc_bb(y=df$y, n=df$n, subj=df$subj, x=cbind(1,df$x), xnull=matrix(1,nrow=nrow(df),ncol=1), niter=200, niter_laplace=2, num.nodes=3,
optim.method="BFGS", converge_tol=1e-8)
str(res.bb)

}
