% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_model.R
\name{generate_desmat}
\alias{generate_desmat}
\title{Generate design matrix that combines genotypes and B-spline basis functions.}
\usage{
generate_desmat(geno.temp, ptbs)
}
\arguments{
\item{geno.temp}{Genotype of cis-SNPs. Matrix of ncells x (number of SNPs).}

\item{ptbs}{B-spline basis functions. Matrix of ncells x (number of B-spline basis functions)}
}
\description{
Generate design matrix that combines genotypes and B-spline basis functions.
}
