% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{twist_association}
\alias{twist_association}
\title{TWiST association analysis}
\usage{
twist_association(
  sumstat,
  wgtlist,
  weights_pred,
  bim_train,
  genos,
  ngwas,
  opt = list(max_impute = 0.6, min_r2pred = 0.8, degree_beta = 3, knots_beta = seq(0.05,
    0.95, by = 0.05), logsigma2vec = seq(-20, 8, by = 0.2))
)
}
\arguments{
\item{sumstat}{A data frame of GWAS summary statistics. Columns: SNP (SNP ID), A1 (effect allele), A2 (other allele), Z (z-statistic for GWAS association).}

\item{wgtlist}{A data frame of gene annotation. Five columns: ID (gene identifier), CHR (chromosome), P0 (start), P1 (end), tss (transcription start site).}

\item{weights_pred}{Named list where each element is an output of twist_train_model, fast_twist_train_model, or fpca_train_model. The number and order of genes should be the same as \code{wgtlist}. Named by \code{wgtlist$ID}.}

\item{bim_train}{bim file from the training data where prediction models were built . Should include all the SNPs in \code{weights_pred}}

\item{genos}{Reference data in plink format used to compute LD. Can be read into R using plink2R. A list of length 3: bed, bim, fam. Usually 1000 Genomes.}

\item{ngwas}{GWAS effective sample size. For continuous traits, it is the total sample size; for case-control binary traits, it is ncases*ncontrols/(ncases+ncontrols). If an integer, assumes all SNPs have the same sample size. Can also be a vector with SNP-specific sample sizes.}

\item{opt}{Options, including \code{max_impute} (maximum proportion of missing SNPs), \code{min_r2pred} (miminum average GWAS Z-score imputation r2), \code{degree_beta} (degree of B-spline basis functions), \code{knots_beta} (degree of B-spline basis functions), \code{logsigma2vec} (candidate log-sigma2 values for grid search).}
}
\value{
A list including
\item{out.tbl}{A data frame containing gene and model information, including the following columns:
ID (gene identifier), CHR (chromosome), P0 (start), P1 (end), tss (transcription start site), sigma2 (variance of coefficients of B-spline basis functions \eqn{\boldsymbol{\beta}}),
p.global (p-value for global test), p.dynamic (p-value for dynamic test), p.nonlinear (p-value for nonlinear test), degree (degree of B-spline basis functions for \eqn{\beta(t)}.}
\item{betal}{Estimated coefficients for linear component of gene-trait effect \eqn{\beta(t)}.}
\item{var.betal}{Variance-covariance matrix of \code{betal}.}
\item{beta}{Estimated coefficients for nonlinear component of gene-trait effect \eqn{\beta(t)}.}
\item{var.beta}{Variance-covariance matrix of \code{beta}.}
\item{knots}{Knots for B-spline basis functions for gene-trait effect \eqn{\beta(t)}.}
}
\description{
This function conducts single-cell TWAS analysis using GWAS summary statistics and weights trained in TWiST Stage-1 model. The output for each gene is an estimated curve representing gene-trait effects along the pseudotime course, as well as p-values for global, dynamic, and nonlinear test.
}
\details{
TWiST association analysis is conducted using scalar-on-function regression, with the genetically regulated expression (denoted by \eqn{v_i(t)}) as the predictor and the trait as the outcome. This functions uses GWAS summary statistics instead of individual-level data.
The estimated gene-trait effect is
\deqn{\beta(t)=\beta_{l0}+\beta_{l01}t+\sum\limits_{m=1}^{M'}\beta_m\psi_m(t),}
where \eqn{(\beta_{l0},\beta_{l01})^T} (\code{betal} in the R code) is the vector coefficients for the linear effects, and \eqn{(\beta_1,...,\beta_{M'})^T} (\code{beta} in the R code) is the vector of coefficients for the nonlinear effects.
}
